feat: Conditionally render base URL page
