+++ backend/routes/authRoutes.js
@@ -3,6 +3,8 @@ const router = express.Router();
 import passport from 'passport';
 import { registerStudent, loginStudent } from '../controllers/studentAuthController.js';
 import { registerTeacher, loginTeacher } from '../controllers/teacherAuthController.js';
+import Student from '../models/Student.js';
+import Teacher from '../models/Teacher.js';
 
 const CLIENT_URL = process.env.FRONTEND_URL || "http://localhost:3000";
 
@@ -32,7 +34,7 @@ router.get(
       maxAge: 3600000,
     });
 
-    res.cookie("name", `${req.user.user.fname} ${req.user.user.lname}`, {
+    res.cookie("email", req.user.email, {
       httpOnly: true,
       secure: process.env.NODE_ENV === "production",
       sameSite: "Strict",
@@ -70,7 +72,14 @@ router.get(
       maxAge: 3600000,
     });
 
-    res.cookie("name", `${req.user.user.fname} ${req.user.user.lname}`, {
+    res.cookie("email", req.user.email, {
+      httpOnly: true,
+      secure: process.env.NODE_ENV === "production",
+      sameSite: "Strict",
+      maxAge: 3600000,
+    });
+
+    res.cookie("profilePicture", req.user.picture, {
       httpOnly: true,
       secure: process.env.NODE_ENV === "production",
       sameSite: "Strict",
@@ -82,23 +91,47 @@ router.get(
   }
 );
 
-router.get("/check", (req, res) => {
+router.get("/check", async (req, res) => {
   console.log("Cookies received:", req.cookies);
 
   if (Object.entries(req.cookies).length > 0) {
-      res.json({ isAuthenticated: true });
+    const Model = req.cookies.role === "student" ? Student : Teacher;
+    
+    try {
+      const user = await Model.findOne({ email: req.cookies.email });
+      console.log("User found:", user);
+
+      if (!user) {
+        return res.json({ isAuthenticated: false });
+      }
+
+      res.json({ 
+        isAuthenticated: true,
+        role: req.cookies.role,
+        email: user.email,
+        token: req.cookies.token,
+        name: `${user.fname} ${user.lname}`,
+        profilePic: req.cookies.profilePicture,
+      });
+    } catch (error) {
+      console.error("Error fetching user:", error);
+      res.status(500).json({ error: "Internal server error" });
+    }
   } else {
-      res.json({ isAuthenticated: false });
+    res.json({ isAuthenticated: false });
   }
 });
+
 
 // Student Auth Routes
 router.post("/student/register", registerStudent);
 router.post("/student/login", loginStudent);
+// router.get("/student/delete", deleteStudent);
 
 // Teacher Auth Routes
 router.post("/teacher/register", registerTeacher);
 router.post("/teacher/login", loginTeacher);
+// router.get("/teacher/delete", deleteTeacher);
 
 router.post("/logout", (req, res) => {
   for(const cookie in req.cookies) {
