+++ frontend/src/app/oauth-success/page.js
@@ -3,8 +3,9 @@ import { useEffect } from "react";
 import { useRouter, useSearchParams } from "next/navigation";
 
 const OAuthSuccess = () => {
-  const router = useRouter();
+  // const router = useRouter();
   const searchParams = useSearchParams();
+  const router = useRouter();
 
   useEffect(() => {
     const token = searchParams.get("token");
@@ -12,7 +13,7 @@ const OAuthSuccess = () => {
 
     if (token) {
       // Redirect to dashboard after login
-      router.push(`/${role}/dashboard`);
+      router.push(`/`);
     } else {
       router.push("/signin");
     }
