+++ frontend/src/utils/auth.js
@@ -3,7 +3,14 @@ import instance from './axios.js';
 export const isAuthenticated = async () => {
     try {
         const response = await instance.get('/api/auth/check', { withCredentials: true });
-        return response.data.isAuthenticated;
+        const data = response.data;
+        return { 
+            isLoggedIn: data.isAuthenticated, 
+            role: data.role,
+            email: data.email,
+            name: data.name,
+            profilePic: data.profilePic,
+        };
     } catch (error) {
         console.log('Auth Check failed:', error);
         return false;
@@ -13,6 +20,8 @@ export const isAuthenticated = async () => {
 export const logout = async () => {
     try {
         await instance.post("/api/auth/logout");
+
+        window.location.href = "/";
     } catch (error) {
         console.error("Logout error:", error);
     }
